Split out UHAgreementHelpers from TenancyClassification::V2::Helper
