+++ lib/hackney/income/tenancy_classification/v2/helpers.rb
@@ -3,6 +3,10 @@ module Hackney
     module TenancyClassification
       module V2
         module Helpers
+          # Replace this with Hackney::Income::TenancyClassification::V2::MAAgreementHelpers
+          # when agreements synced from UH
+          include Hackney::Income::TenancyClassification::V2::Helpers::UHAgreementHelpers
+
           def case_paused?
             @case_priority.paused?
           end
@@ -55,31 +59,6 @@ module Hackney
           def last_communication_newer_than?(date)
             return false if @criteria.last_communication_date.blank?
             @criteria.last_communication_date > date.to_date
-          end
-
-          def active_agreement?
-            @criteria.active_agreement? || (@criteria.most_recent_agreement.present? && @criteria.most_recent_agreement[:status].in?(%i[active breached]))
-          end
-
-          def informal_breached_agreement?
-            return false if should_prevent_action?
-            breached_agreement? && !court_breach_agreement?
-          end
-
-          def breached_agreement?
-            return false if should_prevent_action?
-            return false if @criteria.most_recent_agreement.blank?
-            return false if @criteria.most_recent_agreement[:start_date].blank?
-
-            @criteria.most_recent_agreement[:breached]
-          end
-
-          def court_breach_agreement?
-            return false if should_prevent_action?
-            return false unless breached_agreement?
-            return false if @criteria.courtdate.blank?
-
-            @criteria.most_recent_agreement[:start_date] > @criteria.courtdate
           end
 
           def balance_is_in_arrears_by_number_of_weeks?(weeks)
